gcp/requirements.txt
tensorflow==2.11.0
google-cloud-storage==1.16.1
Pillow==9.4.0
